minimal impl of ck-launch-session behavior in libcros
Sadly, doing this through the normal dbus-glib channels doesn't seem
to work.  I hypothesize that this is because of the myriad of
g_main_loop contexts that get created and destroyed as chrome gets
going.  This seemed to make ConsoleKit believe that the entity that
called "OpenSession" went away, which would de-facto close the
session.  This made us unable to talk to connman.  Anyway, because doing
the dbus-glib stuff didn't work, I went a level down and made pure dbus calls.

Review URL: http://codereview.chromium.org/578001