Add a new member language_code to InputLanguage.
We'll use the language codes for showing input language names in Chromium OS.

The change breaks the backward compatibility (InputLanguage info returned
from libcros to chromium will be incompatible), but there doesn't seem
to be a good way to keep the compatibility as the struct is used in
many places both in libcros and in chromium. So what we can do is to
update cros_deps/DEPS as soon as we check in the change.

Hopefully, we now fetch the latest build of chromium for chromium os
and the latest build is updated frequently, so we don't have to worry
too much about the temporary compatibility breakage.

http://src.chromium.org/cgi-bin/gitweb.cgi?p=chromiumos-overlay.git;a=commit;h=4a34b2232bfda4bf45c45edf3b73ffb1fb90d53e

As far as I tested, a release build of the latest chromium for chromium os
(without cros_deps/DEPS updated) didn't crash with this change as long as
you don't touch the input language selector at the right upper corner.

Review URL: http://codereview.chromium.org/652106