Add cryptohome call to get the system salt.
Dave--are you the appropriate person to take a look at this?  I needed to add a method to get the system salt to remove pam_offline's dependency on being able to read the shadow root directory.  Will this require updating the libcros version number?

Review URL: http://codereview.chromium.org/2012013