Swap D-Bus and IBus initialization order to avoid race conditions.
Initialize D-Bus connection first so that the destructor of the InputMethodStatusConnection class does not terminate the IBus connection between im-ibus.so and ibus-daemon.
This change is not for a specific bug, but might be worth not to trigger subtle bugs in im-ibus, libibus, and libdbus.

BUG=3683
TEST=manually

Review URL: http://codereview.chromium.org/2627009

Conflicts:

	chromeos_input_method.cc