Changed signature of MonitorPropertyCallback.
It now takes a Value as defined in base/values.h instead of a glib::Value.
The definition of the former is available to Chrome. The definition of the
latter is not.

Did not increment CrosAPIVersion, because no one was using
MonitorPropertyCallback yet.

BUG=
TEST=Tested using monitor_network to watch property changes.
Verified that older version of chrome works with this version
of libcros.

Change-Id: I2be4c6663f91cf78b73a8d3caf7b71e8bc8f0324

Review URL: http://codereview.chromium.org/3828002