Change ListIPConfigs to return the hardware address as well.
We'll show the hardware address in the internet options page.

We could instead introduce a new function that returns the full device
info with IP configs attached (this is what flimflam returns), but that
would be a big change and need a lot of refactoring (for instance, it
would be nice to have 1:1 mappings with flimflam APIs and data structures).
We should probably do it at some point after R9 release.

BUG=chromium-os:3648
TEST=Confirmed that the internet options worked as expected with this change.

Change-Id: I0a5da7910c0bb8bca9e7f37fa4b8b5245ec5caf9

Review URL: http://codereview.chromium.org/3916003