Do not send UpdateUI notification to Chrome when focus_in signal is sent from ibus-daemon.
Since every Chrome window issues IBus request on the UpdateUI notification and the total cost of the calls might not be so cheap (especially when there're many Chrome windows), we should avoid sending the notification to Chrome as much as possible.

In the past, updating UI on focus_in was definitely necessary since IME status was per input context. However, since we have already implemented the "global IME" feature, it should be unnecessary now.

BUG=chromium-os:8284
TEST=manually done. (1) on login screen, click US and select INTL, then use IME hotkeys (2) on screen locker, use IME hotkeys (3) on browser, use IME hotkeys (4) on browser, open IME DOMUI and (4-1) disable all IMEs except US and check IME status button ("US") disappears (4-2) disable current global engine and check the IME status button changes.

Change-Id: I5730dabed6eed3eb141141fda9bb1f946b1eef73

Review URL: http://codereview.chromium.org/4206007