Don't try to fetch IPConfigs when the device path passed in is empty.
The ListIPConfigs function was checking for a NULL device path, but
was not checking for a zero-length string. Because chrome converts
the char * into a std::string (creating an empty string if the char *
is NULL), when chrome passed the device path back in, it was passing
an empty string. Since an empty DBus path is not valid, an assertion
was failing, crashing chrome.

Normally, chrome would not see a ServiceInfo structure where the device
path is NULL. But in the disable case, between the time that chrome
gets the list of services from flimflam and the time that chrome asks
for the properties of the cellular service, flimflam may have processed
the disable request to the point where the service no longer has a
device associated with it.

BUG=chromium-os:9790,chromium-os:9840
TEST=Repeatedly disabled 3G under conditions that had been crashing
chrome and verified that the crashes no longer occur.

Change-Id: I68a6dcfcfd164d0109062b10ac0c2961d9009d42

Review URL: http://codereview.chromium.org/5951001