Do not handle the StateChanged D-Bus signal from ibus-daemon.
The signal is sent to the panel when an engine (for an input context) is enabled, disabled, or changed. But since libcros handle the GlobalEngineChanged as well, it can safely ignore the signal.

This should make input method switching faster since more than 75% of UpdateUI() calls in libcros will be removed.

BUG=chromium-os:11908
BUG=chromium-os:9206
TEST=manual

Change-Id: Ibab900289311388b16b9560f1e7d118ac6b30d3e

Review URL: http://codereview.chromium.org/6519008