Fixed an apparent typo that would cause an assertion failure and Chrome crash.
If a property change signal was sent for a flimflam property whose value
was a dictionary, it would try to read the value as the key. Since it is
expecting a string in this context, it triggered an assertion failure.

Also, I changed monitor_network to have the ability to print
dictionary-valued properties. For now, it just handles string -> string
dictionaries.

BUG=none
TEST=Tested with my version of flimflam that makes Cellular.APN a
property with a dictionary for a value. Prior to this fix, setting
that property would result in Chrome crashing. Now it doesn't.

Change-Id: Iccce7e2b7433a40f003347b9a61fdf3ee195b78e

R=njw@chromium.org,stevenjb@chromium.org

Review URL: http://codereview.chromium.org/6675026