Deprecate SetSynapticsParameter which has been unused for quite long time.
The client code in Chrome synaptics_library.h was gone long time ago:

http://src.chromium.org/viewvc/chrome?view=rev&revision=58842

The previous attempt of removing this directly caused chrome not to boot
as cros/load.cc in Chrome tree still tried to load the function. Hence
the change was reverted by davidjames:
http://gerrit.chromium.org/gerrit/#change,3418
We should have followed the right procesure even though the API was
unused for a year.

BUG=chromium-os:16561
TEST=gmerge libcros, and conform chrome boots

Change-Id: I755f6626950d7b50ce9c631a9a0590a3e62bf5da
Reviewed-on: http://gerrit.chromium.org/gerrit/3584
Tested-by: Satoru Takabayashi <satorux@chromium.org>
Reviewed-by: Dave Moore <davemoore@chromium.org>