Adding hidden propertyu to DiskInfo.
The property tells us if the device should be shown in the UI (e.g in
fileBrowser), or mount library should just keep track of it (we need this
information for imageburner to work).

Current use of is_hidden property in the event from cros_disks is relict from
times when DeviceKit was used and has no practical purpose anymore, so we are
reusing it.

Also this CL stops setting partition slave property of DiskInfo, since it is not
used anymore (and is always NULL).

Chrome side of the change: http://codereview.chromium.org/8049002/

TEST=run opt/google/chrome/chromeos/monitor_mount and checked that is_hidden property gets set as it should for inserted device
BUG=chromium-os:20476

Change-Id: I3acfd4e7c20c8fcd467a28cf63563cbf634b85e4
Reviewed-on: http://gerrit.chromium.org/gerrit/8216
Reviewed-by: Zelidrag Hornung <zelidrag@chromium.org>
Tested-by: Toni Barzic <tbarzic@chromium.org>