chromeos_power: read battery state thru dbus as bool
The only BatteryState enum value being used in Chrome is
BATTERY_STATE_FULLY_CHARGED.  Thus, as part of eliminating libcros,
powerd now only sends a boolean indicating whether the battery is fully
charged.  This way, there is no need to share an enum definition between
Chrome and Chrome OS.

This patch still keeps the BatteryState enum and converts the boolean
back to it, as a temporary stopgap.  Eventually the code in
chromeos_power will be moved to Chrome.

BUG=chromium-os:16566
TEST=none

Change-Id: I66d8b272649431a15363878f60a31a4a99897963
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/8250
Reviewed-by: Satoru Takabayashi <satorux@chromium.org>