Do not attempt to receive a return value from EmitLoginPromptVisible.
Because the D-Bus method EmitLoginPromptVisible does not return a value.
See platform/login_manager/session_manager.xml for the signature.

Return 'true' from chromeos::EmitLoginPromptVisible() instead of making the
return value type to 'void' , as changing the libcros API is inherently
tricky. This is fine, as the return value is not used in Chrome anyway.

TEST=libcros builds
BUG=chromium-os:20517

Change-Id: Ib5b7db49184ea32b1fb99e67d14f15ac81d154df
Reviewed-on: http://gerrit.chromium.org/gerrit/10225
Tested-by: Satoru Takabayashi <satorux@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>